Vga: Fix init_back to be consistent across the interface.

+++ include/vga_base.h
@@ -63,7 +63,7 @@ public:
  
 	virtual int    init() =0;
 	virtual int    init_front(VgaBuf*) =0;
-	virtual int    init_back(VgaBuf*, DWORD =0, DWORD =0) =0;
+	virtual int    init_back(VgaBuf*, unsigned long =0, unsigned long =0) =0;
 	virtual void   deinit() =0;
 
 	virtual char   is_inited() =0;
